Split manifest and status reporting out of main convention code
