+++ source/Calamari/Kubernetes/Conventions/HelmUpgradeWithKOSConvention.cs
@@ -17,11 +17,9 @@ using Calamari.Deployment.Conventions;
 using Calamari.Kubernetes.Helm;
 using Calamari.Kubernetes.Integration;
 using Calamari.Kubernetes.ResourceStatus;
-using Calamari.Kubernetes.ResourceStatus.Resources;
 using Calamari.Util;
 using Newtonsoft.Json;
 using Octopus.CoreUtilities.Extensions;
-using YamlDotNet.RepresentationModel;
 
 namespace Calamari.Kubernetes.Conventions
 {
@@ -54,8 +52,6 @@ namespace Calamari.Kubernetes.Conventions
 
         public void Install(RunningDeployment deployment)
         {
-            var isKOSEnabled = deployment.Variables.GetFlag(SpecialVariables.ResourceStatusCheck);
-
             var releaseName = GetReleaseName(deployment.Variables);
 
             var helmCli = new HelmCli(log, commandLineRunner, deployment);
@@ -73,21 +69,17 @@ namespace Calamari.Kubernetes.Conventions
             var helmUpgradeTask = Task.Run(() => ExecuteHelmUpgrade(deployment,
                                                                     helmCli,
                                                                     releaseName,
-                                                                    isKOSEnabled,
                                                                     kosCts));
 
             var manifestAndStatusCheckTask = Task.Run(async () =>
                                                       {
-                                                          var manifest = await PollForManifest(deployment, helmCli, releaseName, newRevisionNumber);
+                                                          var runner = new HelmManifestAndStatusReporter(log, statusReporter, manifestReporter);
 
-                                                          //report the manifest has been applied
-                                                          manifestReporter.ReportManifestApplied(manifest);
-
-                                                          //if resource status (KOS) is enabled, parse the manifest and start monitored the resources
-                                                          if (isKOSEnabled)
-                                                          {
-                                                              await ParseManifestAndMonitorResourceStatuses(deployment, manifest, kosCts.Token);
-                                                          }
+                                                          await runner.StartMonitoringManifestAndReportingStatus(deployment,
+                                                                               helmCli,
+                                                                               releaseName,
+                                                                               newRevisionNumber,
+                                                                               kosCts.Token);
                                                       },
                                                       kosCts.Token);
 
@@ -95,91 +87,14 @@ namespace Calamari.Kubernetes.Conventions
             Task.WhenAll(helmUpgradeTask, manifestAndStatusCheckTask).GetAwaiter().GetResult();
         }
 
-        async Task<string> PollForManifest(RunningDeployment deployment,
-                                           HelmCli helmCli,
-                                           string releaseName,
-                                           int revisionNumber)
-        {
-            var ct = new CancellationTokenSource();
-            var timeout = GoDurationParser.TryParseDuration(deployment.Variables.Get(SpecialVariables.Helm.Timeout), out var timespan) ? timespan : TimeSpan.FromMinutes(5);
-            ct.CancelAfter(timeout);
-            string manifest = null;
-            while (!ct.IsCancellationRequested)
-            {
-                try
-                {
-                    manifest = helmCli.GetManifest(releaseName, revisionNumber);
-                    log.Verbose($"Retrieved manifest for {releaseName}, revision {revisionNumber}.");
-                    break;
-                }
-                catch (CommandLineException)
-                {
-                    log.Verbose("Helm manifest was not ready for retrieval. Retrying in 1s.");
-                    await Task.Delay(TimeSpan.FromSeconds(1), ct.Token);
-                }
-            }
-
-            if (string.IsNullOrWhiteSpace(manifest))
-            {
-                throw new CommandException("Failed to retrieve helm manifest in a timely manner");
-            }
-
-            return manifest;
-        }
-
-        async Task ParseManifestAndMonitorResourceStatuses(RunningDeployment deployment, string manifest, CancellationToken cancellationToken)
-        {
-            var resources = new List<ResourceIdentifier>();
-            using (var reader = new StringReader(manifest))
-            {
-                var yamlStream = new YamlStream();
-                yamlStream.Load(reader);
-
-                foreach (var document in yamlStream.Documents)
-                {
-                    if (!(document.RootNode is YamlMappingNode rootNode))
-                    {
-                        log.Warn("Could not parse manifest, resources will not be added to kubernetes object status");
-                        continue;
-                    }
-
-                    var kind = rootNode.GetChildNode<YamlScalarNode>("kind").Value;
-
-                    var metadataNode = rootNode.GetChildNode<YamlMappingNode>("metadata");
-                    var name = metadataNode.GetChildNode<YamlScalarNode>("name").Value;
-                    var @namespace = metadataNode.GetChildNodeIfExists<YamlScalarNode>("namespace")?.Value;
-
-                    //if the resource doesn't have a namespace set in the manifest set it to the helm namespace.
-                    //This is because namespaced resources that don't have the namespace defined in the manifest will be set in the namespace set in the helm command
-                    //if this is null, we'll fall back on the namespace defined for the kubectl tool (which is the default target namespace)
-                    //we aren't changing the manifest here, just changing where the kubectl looks for our resource.
-                    //We also try and filter out known non-namespaced resources
-                    if (string.IsNullOrWhiteSpace(@namespace) && !KubernetesApiResources.NonNamespacedKinds.Contains(kind))
-                    {
-                        @namespace = deployment.Variables.Get(SpecialVariables.Helm.Namespace)?.Trim();
-                    }
-
-                    var resourceIdentifier = new ResourceIdentifier(kind, name, @namespace);
-                    resources.Add(resourceIdentifier);
-                }
-            }
-
-            //We are using helm as the deployment verification so an infinite timeout and wait for jobs makes sense
-            var statusCheck = statusReporter.Start(0, false, resources);
-            await statusCheck.WaitForCompletionOrTimeout(cancellationToken);
-        }
-
         void ExecuteHelmUpgrade(RunningDeployment deployment,
                                 HelmCli helmCli,
                                 string releaseName,
-                                bool isKOSEnabled,
                                 CancellationTokenSource cts)
         {
             var packagePath = GetChartLocation(deployment);
 
-            var args = GetUpgradeCommandArgs(deployment,
-                                             helmCli,
-                                             isKOSEnabled);
+            var args = GetUpgradeCommandArgs(deployment, helmCli);
 
             var result = helmCli.Upgrade(releaseName, packagePath, args);
 
@@ -201,9 +116,7 @@ namespace Calamari.Kubernetes.Conventions
             }
         }
 
-        List<string> GetUpgradeCommandArgs(RunningDeployment deployment,
-                                           HelmCli helmCli,
-                                           bool isKOSEnabled)
+        List<string> GetUpgradeCommandArgs(RunningDeployment deployment, HelmCli helmCli)
         {
             var args = new List<string>();
 
@@ -234,7 +147,7 @@ namespace Calamari.Kubernetes.Conventions
             var hasAdditionalArgs = SetAdditionalArguments(deployment, args);
 
             //Adjust args based on KOS
-            if (isKOSEnabled)
+            if (deployment.Variables.GetFlag(SpecialVariables.ResourceStatusCheck))
             {
                 AddKOSArgs(deployment.Variables, hasAdditionalArgs, args);
             }
@@ -372,13 +285,6 @@ namespace Calamari.Kubernetes.Conventions
 
             args.Add($"--tiller-connection-timeout \"{tillerTimeout}\"");
         }
-
-        static string SyntaxSpecificFileName(RunningDeployment deployment, ScriptSyntax syntax)
-        {
-            return Path.Combine(deployment.CurrentDirectory, $"Calamari.HelmUpgrade.{SyntaxSpecificFileExtension(syntax)}");
-        }
-
-        static string SyntaxSpecificFileExtension(ScriptSyntax syntax) => syntax == ScriptSyntax.PowerShell ? "ps1" : "sh";
 
         string GetReleaseName(IVariables variables)
         {
